T132029385 add all merge tests in one file
