+++ faiss/IndexIVF.h
@@ -302,14 +302,8 @@ struct IndexIVF : Index, Level1Quantizer {
 
     size_t remove_ids(const IDSelector& sel) override;
 
-    /** check that the two indexes are compatible (ie, they are
-     * trained in the same way and have the same
-     * parameters). Otherwise throw. */
     void check_compatible_for_merge(const Index& otherIndex) const override;
 
-    /** moves the entries from another dataset to self. On output,
-     * other is empty. add_id is added to all moved ids (for
-     * sequential ids, this would be this->ntotal */
     virtual void merge_from(Index& otherIndex, idx_t add_id) override;
 
     /** copy a subset of the entries index to the other index
